requirements.txt
numpy
scipy
llama-cpp-python==0.1.77
sounddevice
piper-tts
Levenshtein
pysdl2
pysdl2-dll
soundfile
fasteners
colorama
psutil
ctranslate2
transformers
pasimple
pygame_menu
torch
torchaudio
pyserial
